half precision instead of full precision

+++ all/memory/replay_buffer.py
@@ -33,7 +33,7 @@ class ExperienceReplayBuffer(ReplayBuffer):
     def sample(self, batch_size):
         keys = np.random.choice(len(self.buffer), batch_size, replace=True)
         minibatch = [self.buffer[key] for key in keys]
-        return self._reshape(minibatch, torch.ones(batch_size, device=self.device))
+        return self._reshape(minibatch, torch.ones(batch_size, device=self.device, dtype=torch.half))
 
     def update_priorities(self, td_errors):
         pass
@@ -49,7 +49,7 @@ class ExperienceReplayBuffer(ReplayBuffer):
         states = [sample[0] for sample in minibatch]
         actions = [sample[1] for sample in minibatch]
         next_states = [sample[2] for sample in minibatch]
-        rewards = torch.tensor([sample[3] for sample in minibatch], device=self.device).float()
+        rewards = torch.tensor([sample[3] for sample in minibatch], device=self.device, dtype=torch.half)
         return (states, actions, next_states, rewards, weights)
 
     def __len__(self):
@@ -107,7 +107,7 @@ class PrioritizedReplayBuffer(ExperienceReplayBuffer):
         weights = np.array(weights, dtype=np.float32)
         samples = [self.buffer[idx] for idx in idxes]
         self._cache = idxes
-        return self._reshape(samples, torch.from_numpy(weights).to(self.device))
+        return self._reshape(samples, torch.from_numpy(weights).half().to(self.device))
 
     def update_priorities(self, td_errors):
         idxes = self._cache
